Remove test discovery flag from `TestExplorer`
